ENH: change the procedure for provisioning of study/series UIDs
Given that for the UW-CIRP dataset there are 2 timepoints and 2 modalities
for each subject, assign distinct Study UIDs for distinct combinations of
PatientID and StudyDate, and distinct Series UIDs for distinct combinations
of PatientID, StudyDate and Modality. Delegate this to a helper class.

Prior to this change, assignment of Study/Series UIDs was problematic.